Add Input Error Handling around zip codes
## Front End

Basic form input validation was added to only allow numeric input with some value present.

## Controller

We now avoid making a call to the API if somehow the zip code is not present in the request.  In addition, an alert is presented for the user indicating that nothing was found.

## API Service

The API service and view now now detect and render feedback when the external service does not recognize the zip-code (or any parameters that have been provided for that matter).